pyproject.toml
[project]
name = "genai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "google>=3.0.0",
    "google-genai>=1.65.0",
    "ipykernel>=7.2.0",
    "ollama>=0.6.1",
    "openai>=2.24.0",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
]
